Add rollbar to log
A request was put through to log the messages from the distributer to
rollbar. This was the simplest way to implement it. The use case is that
logging using Rollbar means that you log straight to a online server which
automatically sends notifications via email when errors take place
together with stack traces